pyproject.toml
[project]
name = "deep-portfolio-theory"
version = "0.1.0"
description = "Deep Portfolio Theory"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "torch",
    "matplotlib",
    "tqdm",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "ipython"]
